:scissors: refactor(app): move settings to radar view
