+++ app/src/components/radar/Blip.vue
