+++ app/src/router/index.ts
@@ -2,7 +2,6 @@ import Vue from 'vue'
 import Router from 'vue-router'
 import Radar from '@/components/Radar.vue'
 import Home from '@/components/Home.vue'
-import List from '@/components/List.vue'
 import ErrorPage from '@/components/ErrorPage.vue'
 import Logout from '@/components/Logout.vue'
 import Users from '@/components/Users.vue'
@@ -13,7 +12,6 @@ Vue.use(Router)
 const routesCfg = appConfig.routes
 const components = {
   home: Home,
-  list: List,
   logout: Logout,
   users: Users,
   radar: Radar,
